DATA-2860 add backfill pipeline (#13)
DATA-2860 This feature adds a new pipeline that is intended for backfills from BQ.backup. 
It reuses pipeline code from dynamodb/salesforce but instead of taking a pub/sub message it takes in a query to extract BQ data with in batch mode. It also writes to BQ instead of passing to pub/sub as with old backfill pipe. 
---------
Co-authored-by: jbergenblom <jonathan.bergenblom@mathem.se>
Co-authored-by: joekgu <johan.gunnarsson@mathem.se>